chore: cleanup handling of async findings reload in tests We're now avoiding a return value of FindingsManager.reloadAsync() in favor of the message sent on the project's message bus.
