+++ plugin-core/src/test/java/appland/problemsView/FindingsManagerTest.java
@@ -7,9 +7,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.util.List;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
 public class FindingsManagerTest extends AppMapBaseTest {
     @Before
@@ -32,17 +30,18 @@ public class FindingsManagerTest extends AppMapBaseTest {
     }
 
     @Test
-    public void findingsVSCodeSystem() throws ExecutionException, InterruptedException, TimeoutException {
+    public void findingsVSCodeSystem() throws Exception {
         var manager = FindingsManager.getInstance(getProject());
+
+        var condition = TestFindingsManager.createFindingsCondition(getProject(), getTestRootDisposable());
         var root = myFixture.copyDirectoryToProject("vscode/workspaces/project-system", "root");
+        assertTrue(condition.await(30, TimeUnit.SECONDS));
 
         var findingsFile = root.findChild("appmap-findings.json");
         assertNotNull(findingsFile);
 
         var problematicFile = root.findFileByRelativePath("app/controllers/microposts_controller.rb");
         assertNotNull(problematicFile);
-
-        manager.reloadAsync().blockingGet(30, TimeUnit.SECONDS);
 
         assertEquals(1, manager.getProblemFileCount());
         assertEquals(1, manager.getProblemCount());
